0-created an async generator function

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -7,7 +7,7 @@ import asyncio
 import random
 
 
-async def async_generator() -> typing.AsyncGenerator[int, None]:
+async def async_generator() -> typing.AsyncGenerator[float, None]:
     """
     Yields numbers
     """
